# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vue Calculator
  description: |-
    Just a simple calculator
  version: 1.0.0
servers:
  - url: localhost:3000
...
paths:
...
  /user/token:
    description: Login users or get currently logged in user
    post:
      tags:
        - Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "201":
          description: Successfully logged in the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenData"
        "401":
          description: Could not log user in. Wrong credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "404":
          description: The user does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "500":
          description: Uncaught internal server error
    put:
      description: Create a new JWT token given a previous valid one.
      tags:
        - Token
      responses:
        "201":
          description: Successfully created a new JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenData"
        "404":
          description: There is no current logged in session to refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
        "500":
          description: Uncaught internal server error
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          examples:
            - johndoe
        password:
          type: string
          examples:
            - supersecure123
...
    TokenData:
      type: object
      description: The request body for requesting a new token given another valid token
      required:
        - userId
        - jwtToken
      properties:
        userId:
          type: string
        jwtToken:
          type: string
    GenericError:
      type: object
      description: A generic error message response
      required:
        - message
      properties:
        message:
          type: string